Fix delete errors and refactor integration tests
- Remove error logging for 'not found' on delete in all controllers
  (Agent, MCPServer, ModelAPI). These were causing false errors when
  resources were deleted, since reconcile is called after deletion.

- Refactor integration tests into separate files by controller:
  - agent_test.go (6 tests)
  - mcpserver_test.go (4 tests)
  - modelapi_test.go (6 tests)

- Add delete tests to verify no errors occur during CRD deletion
  - Tests verify finalizers are properly removed
  - Note: envtest doesn't run garbage collection

Total: 15 integration tests (was 12)
